Added SplashScreen, sourced out Username
