+++ app/src/main/java/com/example/ss18/msp/lmu/msp_projectkickoff_ss188/Activities/MainActivity.java
@@ -34,50 +34,20 @@ import com.example.ss18.msp.lmu.msp_projectkickoff_ss188.Users.User;
 
 public class MainActivity extends AppCompatActivity {
 
-    private static final String PREFS_NAME = "preferences_title";
-    private static final String PREF_USER = "preferences_username";
+
 
     private ImageButton presenter;
     private ImageButton spectator;
-    private AlertDialog alertDialog;
 
-    /**
-     * ACCESS_COARSE_LOCATION is considered dangerous, so we need to explicitly
-     * grant the permission every time we start the app
-     */
-    private static final String[] REQUIRED_PERMISSIONS =
-            new String[]{
-                    Manifest.permission.ACCESS_COARSE_LOCATION,
-                    Manifest.permission.BLUETOOTH,
-                    Manifest.permission.BLUETOOTH_ADMIN,
-                    Manifest.permission.ACCESS_WIFI_STATE,
-                    Manifest.permission.CHANGE_WIFI_STATE,
-            };
+
 
     /**
      * Tag for Logging/Debugging
      */
     private static final String TAG = "MAIN_ACTIVITY";
 
-    private boolean userNameAlreadyEntered = false;
 
-    /**
-     * Called when our Activity has been made visible to the user.
-     * This is only needed for newer devices
-     */
-    @RequiresApi(api = Build.VERSION_CODES.M)
-    @Override
-    protected void onStart() {
-        super.onStart();
-        //Check if we have all permissions, if not, then add!
-        for (String permission : REQUIRED_PERMISSIONS) {
-            if (ContextCompat.checkSelfPermission(this, permission)
-                    != PackageManager.PERMISSION_GRANTED) {
-                requestPermissions(REQUIRED_PERMISSIONS, 1);
-                return;
-            }
-        }
-    }
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -87,13 +57,6 @@ public class MainActivity extends AppCompatActivity {
         //Views
         presenter = findViewById(R.id.buttonPresenter);
         spectator = findViewById(R.id.buttonSpectator);
-        //Setting up username via AlertDialog
-        loadPreferences();
-        if (!userNameAlreadyEntered) {
-            setUsername();
-        } else {
-            Toast.makeText(MainActivity.this, "HI! " + LocalDataBase.getUserName(), Toast.LENGTH_LONG).show();
-        }
 
         //Animations
         presenter.startAnimation(logoMoveAnimation);
@@ -171,87 +134,11 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        if(alertDialog != null) alertDialog.dismiss();
     }
 
-    /**
-     * Saves the username to the preferences
-     */
-    private void savePreferences() {
-        SharedPreferences settings = getSharedPreferences(PREFS_NAME,
-                Context.MODE_PRIVATE);
-        SharedPreferences.Editor editor = settings.edit();
 
-        // Edit and commit
-        Log.i(TAG,"Save username: " + LocalDataBase.getUserName());
-        editor.putString(PREF_USER, LocalDataBase.getUserName());
-        editor.commit();
-    }
 
-    /**
-     * Loads the username form the preferences
-     */
-    private void loadPreferences() {
-        SharedPreferences settings = getSharedPreferences(PREFS_NAME,
-                Context.MODE_PRIVATE);
-        // Set username if already existing
-        if(userNameAlreadyEntered = settings.contains(PREF_USER))
-        {
-            LocalDataBase.setUserName(settings.getString(PREF_USER,LocalDataBase.getUserName()));
-            Log.i(TAG,"Load username: " + LocalDataBase.getUserName());
-        }
-    }
 
-    /**
-     * Displays an input dialog window to enter the username
-     */
-    private void setUsername() {
-        final EditText input = new EditText(this);
-        AlertDialog.Builder dialog = new AlertDialog.Builder(this);
-        dialog.setTitle(R.string.welcome);
-        dialog.setMessage(R.string.username);
-        dialog.setView(input);
-        dialog.setPositiveButton(R.string.enter, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                LocalDataBase.setUserName(input.getText().toString());
-                savePreferences();
-                Toast.makeText(MainActivity.this, "Welcome " +
-                        LocalDataBase.getUserName()+ "!", Toast.LENGTH_LONG).show();
-            }
-        });
-        dialog.setNegativeButton(R.string.skip, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                Toast.makeText(MainActivity.this, "Skipped entering a username", Toast.LENGTH_LONG).show();
-                savePreferences();
-            }
-        });
-
-        alertDialog = dialog.create();
-        alertDialog.show();
-    }
-
-    /**
-     * Called when the user has accepted (or denied) our permission request.
-     */
-    @CallSuper
-    @Override
-    public void onRequestPermissionsResult(
-            int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (requestCode == 1) {
-            for (int grantResult : grantResults) {
-                if (grantResult == PackageManager.PERMISSION_DENIED) {
-                    Toast.makeText(this, R.string.missingPermission, Toast.LENGTH_LONG).show();
-                    Log.e(TAG, "Permission error");
-                    finish();
-                    return;
-                }
-            }
-            recreate();
-        }
-    }
 
     /**
      * Gets executed if the user chooses to be a "Presenter"  by pressing
