+++ app/src/main/java/com/example/ss18/msp/lmu/msp_projectkickoff_ss188/Activities/SplashActivity.java
@@ -1,0 +1,129 @@
+package com.example.ss18.msp.lmu.msp_projectkickoff_ss188.Activities;
+
+import android.Manifest;
+import android.content.Context;
+import android.content.DialogInterface;
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
+import android.os.Build;
+import android.os.Bundle;
+import android.os.PersistableBundle;
+import android.support.annotation.CallSuper;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.annotation.RequiresApi;
+import android.support.v4.content.ContextCompat;
+import android.support.v7.app.AlertDialog;
+import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
+import android.widget.EditText;
+import android.widget.Toast;
+
+import com.example.ss18.msp.lmu.msp_projectkickoff_ss188.DataBase.LocalDataBase;
+import com.example.ss18.msp.lmu.msp_projectkickoff_ss188.R;
+
+public class SplashActivity extends AppCompatActivity {
+    private static final String PREFS_NAME = "preferences_title";
+    private static final String PREF_USER = "preferences_username";
+
+    private boolean userNameAlreadyEntered = false;
+    private AlertDialog alertDialog;
+
+    /**
+     * Tag for Logging/Debugging
+     */
+    private static final String TAG = "SPLASH_ACTIVITY";
+
+
+    /**
+     * ACCESS_COARSE_LOCATION is considered dangerous, so we need to explicitly
+     * grant the permission every time we start the app
+     */
+    private static final String[] REQUIRED_PERMISSIONS =
+            new String[]{
+                    Manifest.permission.ACCESS_COARSE_LOCATION,
+                    Manifest.permission.BLUETOOTH,
+                    Manifest.permission.BLUETOOTH_ADMIN,
+                    Manifest.permission.ACCESS_WIFI_STATE,
+                    Manifest.permission.CHANGE_WIFI_STATE,
+            };
+    /**
+     * Called when our Activity has been made visible to the user.
+     * This is only needed for newer devices
+     */
+    @RequiresApi(api = Build.VERSION_CODES.M)
+    @Override
+    protected void onStart() {
+        super.onStart();
+        //Check if we have all permissions, if not, then add!
+        for (String permission : REQUIRED_PERMISSIONS) {
+            if (ContextCompat.checkSelfPermission(this, permission)
+                    != PackageManager.PERMISSION_GRANTED) {
+                requestPermissions(REQUIRED_PERMISSIONS, 1);
+                return;
+            }
+        }
+    }
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+        if(alertDialog != null) alertDialog.dismiss();
+    }
+
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        setTheme(R.style.AppTheme);
+        super.onCreate(savedInstanceState);
+
+        loadPreferences();
+        if (!userNameAlreadyEntered) {
+            Intent intent = new Intent(this,RegisterActivity.class);
+            startActivity(intent);
+        } else {
+            //Toast.makeText(SplashActivity.this, "HI! " + LocalDataBase.getUserName(), Toast.LENGTH_LONG).show();
+            Intent intent = new Intent(this, MainActivity.class);
+            startActivity(intent);
+        }
+    }
+
+    /**
+     * Loads the username form the preferences
+     */
+    private void loadPreferences() {
+        SharedPreferences settings = getSharedPreferences(PREFS_NAME,
+                Context.MODE_PRIVATE);
+        // Set username if already existing
+        if(userNameAlreadyEntered = settings.contains(PREF_USER))
+        {
+            LocalDataBase.setUserName(settings.getString(PREF_USER,LocalDataBase.getUserName()));
+            Log.i(TAG,"Load username: " + LocalDataBase.getUserName());
+        }
+    }
+
+
+
+
+
+    /**
+     * Called when the user has accepted (or denied) our permission request.
+     */
+    @CallSuper
+    @Override
+    public void onRequestPermissionsResult(
+            int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if (requestCode == 1) {
+            for (int grantResult : grantResults) {
+                if (grantResult == PackageManager.PERMISSION_DENIED) {
+                    Toast.makeText(this, R.string.missingPermission, Toast.LENGTH_LONG).show();
+                    Log.e(TAG, "Permission error");
+                    finish();
+                    return;
+                }
+            }
+            recreate();
+        }
+    }
+}
