+++ app/src/main/res/layout/activity_main.xml
@@ -29,10 +29,6 @@
         android:layout_height="73dp"
         android:gravity="center"
         android:text="@string/introInfo"
-        android:textColor="#FFFF"
-        android:textSize="24sp"
-        android:textStyle="bold"
-        app:fontFamily="@font/capriola"
         app:layout_constraintBottom_toTopOf="@+id/buttonSpectator"
         app:layout_constraintEnd_toEndOf="@+id/buttonSpectator"
         app:layout_constraintStart_toStartOf="@+id/buttonSpectator"
