+++ app/src/main/res/layout/inbox_fragment.xml
@@ -26,11 +26,8 @@
             android:layout_marginEnd="64dp"
             android:layout_marginStart="84dp"
             android:layout_marginTop="14dp"
-            android:fontFamily="@font/capriola"
             android:text="Inbox"
             android:textAlignment="center"
-            android:textSize="40dp"
-            android:textStyle="bold"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.5"
             app:layout_constraintStart_toStartOf="parent"
