+++ app/src/main/res/layout/participants_fragment.xml
@@ -29,9 +29,7 @@
             android:text="@string/participants"
             android:textColor="@color/colorPrimaryLight"
             android:textSize="40dp"
-            android:onClick="manageParticipants"
-            android:textStyle="bold"
-            app:fontFamily="@font/capriola" />
+            android:onClick="manageParticipants"/>
 
         <TextView
             android:id="@+id/numberOfParticipants"
@@ -43,8 +41,6 @@
             android:text="None :("
             android:textAlignment="center"
             android:textAllCaps="false"
-            android:textSize="90sp"
-            app:fontFamily="@font/capriola"
             tools:layout_editor_absoluteY="66dp" />
 
         <Button
@@ -54,10 +50,7 @@
             android:background="@color/colorPrimaryDark"
             android:elevation="15dp"
             android:text="@string/poke"
-            android:textColor="@color/colorPrimaryLight"
-            android:textSize="40dp"
-            android:textStyle="bold"
-            app:fontFamily="@font/capriola" />
+            android:textColor="@color/colorPrimaryLight"/>
 
         <Button
             android:id="@+id/callButton"
@@ -66,10 +59,7 @@
             android:background="@color/colorAccent"
             android:elevation="15dp"
             android:text="@string/call"
-            android:textColor="@color/colorPrimaryLight"
-            android:textSize="40dp"
-            android:textStyle="bold"
-            app:fontFamily="@font/capriola" />
+            android:textColor="@color/colorPrimaryLight" />
 
     </LinearLayout>
