+++ app/src/main/res/layout/share_fragment.xml
@@ -34,11 +34,8 @@
         android:layout_marginBottom="160dp"
         android:layout_marginEnd="92dp"
         android:layout_marginStart="92dp"
-        android:fontFamily="@font/capriola"
         android:text="Select a file"
         android:textAlignment="center"
-        android:textSize="30sp"
-        android:textStyle="bold"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
