+++ app/src/main/res/values/styles.xml
@@ -8,4 +8,8 @@
         <item name="colorAccent">@color/colorAccent</item>
     </style>
 
+    <style name="AppTheme.Launcher">
+        <item name="android:windowBackground">@drawable/screen_splash</item>
+    </style>
+
 </resources>
